# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Base scaffolding for Go APIs with Gin framework",
        "title": "Simple Arq Golang API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/tiers": {
            "get": {
                "description": "Get all active tiers. Si se pasa role_id, devuelve solo los tiers de ese rol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Get all tiers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filtrar por rol (opcional)",
                        "name": "role_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.TierResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new tier associated with a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Create a new tier",
                "parameters": [
                    {
                        "description": "Tier data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.CreateTierRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.TierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/tiers/by-name": {
            "get": {
                "description": "Get a tier by its name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Get tier by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tier name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.TierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/tiers/{id}": {
            "get": {
                "description": "Get a tier by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Get tier by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.TierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a tier by ID with optional fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Update a tier",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.UpdateTierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.TierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft deletes a tier by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Delete a tier",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_tier.DeleteTierResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "simple-arq-golang_cmd_api_domains_apierror.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "simple-arq-golang_cmd_api_domains_tier.CreateTierRequest": {
            "type": "object",
            "required": [
                "name",
                "role_id"
            ],
            "properties": {
                "description": {
                    "description": "Descripción opcional del tier",
                    "type": "string"
                },
                "name": {
                    "description": "Nombre del tier (requerido)",
                    "type": "string"
                },
                "payment_required": {
                    "description": "Indica si el tier requiere pago",
                    "type": "boolean"
                },
                "role_id": {
                    "description": "ID del rol al que pertenece (requerido)",
                    "type": "integer"
                },
                "tier_amount": {
                    "description": "Monto del tier si requiere pago",
                    "type": "number"
                }
            }
        },
        "simple-arq-golang_cmd_api_domains_tier.DeleteTierResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensaje de confirmación",
                    "type": "string"
                }
            }
        },
        "simple-arq-golang_cmd_api_domains_tier.TierResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Fecha de creación",
                    "type": "string"
                },
                "description": {
                    "description": "Descripción del tier",
                    "type": "string"
                },
                "hierarchy": {
                    "description": "Orden jerárquico (base=1, medium=2, premium=3)",
                    "type": "integer"
                },
                "id": {
                    "description": "ID del tier",
                    "type": "integer"
                },
                "name": {
                    "description": "Nombre del tier",
                    "type": "string"
                },
                "payment_required": {
                    "description": "Indica si el tier requiere pago",
                    "type": "boolean"
                },
                "role_id": {
                    "description": "ID del rol asociado",
                    "type": "integer"
                },
                "role_name": {
                    "description": "Nombre del rol asociado",
                    "type": "string"
                },
                "tier_amount": {
                    "description": "Monto del tier si requiere pago",
                    "type": "number"
                },
                "updated_at": {
                    "description": "Fecha de última actualización",
                    "type": "string"
                }
            }
        },
        "simple-arq-golang_cmd_api_domains_tier.UpdateTierRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Nueva descripción (opcional)",
                    "type": "string"
                },
                "name": {
                    "description": "Nuevo nombre del tier (opcional)",
                    "type": "string"
                },
                "payment_required": {
                    "description": "Indica si el tier requiere pago (opcional)",
                    "type": "boolean"
                },
                "tier_amount": {
                    "description": "Monto del tier (opcional)",
                    "type": "number"
                }
            }
        },
...
    }
}